+++ packages/perseus/src/widgets/__stories__/expression.stories.tsx
@@ -46,7 +46,6 @@ const WrappedKeypadContext = ({
                             apiOptions={{
                                 isMobile: isMobile,
                                 customKeypad: customKeypad,
-                                useV2Keypad: true,
                             }}
                         />
                     );