+++ packages/perseus/src/widgets/__tests__/expression.test.tsx
@@ -42,7 +42,7 @@ const assertCorrect = (itemData: PerseusItem, input: string) => {
         type: "perseus:expression-evaluated",
         payload: {
             result: "correct",
-            virtualKeypadVersion: "PERSEUS_MATH_INPUT",
+            virtualKeypadVersion: "MATH_INPUT_KEYPAD_V2",
         },
     });
 };
@@ -54,7 +54,7 @@ const assertIncorrect = (itemData: PerseusItem, input: string) => {
         type: "perseus:expression-evaluated",
         payload: {
             result: "incorrect",
-            virtualKeypadVersion: "PERSEUS_MATH_INPUT",
+            virtualKeypadVersion: "MATH_INPUT_KEYPAD_V2",
         },
     });
 };
@@ -224,19 +224,8 @@ describe("Expression Widget", function () {
             );
         });
 
-        it("should set the virtual keypad version to MATH_INPUT_KEYPAD_V1 when customKeypad is set and useV2Keypad is unset", () => {
-            assertKeypadVersion({customKeypad: true}, "MATH_INPUT_KEYPAD_V1");
-        });
-
-        it("should set the virtual keypad version to MATH_INPUT_KEYPAD_V2 when customKeypad is set and useV2Keypad is true", () => {
-            assertKeypadVersion(
-                {customKeypad: true, useV2Keypad: true},
-                "MATH_INPUT_KEYPAD_V2",
-            );
-        });
-
-        it("should default the virtual keypad version to PERSEUS_MATH_INPUT", () => {
-            assertKeypadVersion(Object.freeze({}), "PERSEUS_MATH_INPUT");
+        it("should default the virtual keypad version to MATH_INPUT_KEYPAD_V2", () => {
+            assertKeypadVersion(Object.freeze({}), "MATH_INPUT_KEYPAD_V2");
         });
     });
 });